- Moved to Cairo 1.0.4
svn path=/trunk/libgdiplus/; revision=58747

+++ cairo/test/filter-nearest-offset.c
@@ -44,7 +44,7 @@ static cairo_test_status_t
 draw (cairo_t *cr, int width, int height)
 {
     cairo_surface_t *surface;
-    unsigned long data[STAMP_WIDTH * STAMP_HEIGHT] = {
+    uint32_t data[STAMP_WIDTH * STAMP_HEIGHT] = {
 	0xffffffff, 0xffffffff,		0xffff0000, 0xffff0000,
 	0xffffffff, 0xffffffff,		0xffff0000, 0xffff0000,
